+++ db/migrate/001_add_autowatch_involved_issue_to_user.rb
@@ -1,4 +1,4 @@
-class AddAutowatchInvolvedIssueToUser < ActiveRecord::Migration
+class AddAutowatchInvolvedIssueToUser < Rails.version < '5.2' ? ActiveRecord::Migration : ActiveRecord::Migration[4.2]
   def change
     add_column :user_preferences, :autowatch_involved_issue, :boolean, default: true, null: false
   end