+++ test/functional/account_controller_test.rb
@@ -5,18 +5,14 @@ class AccountControllerTest < Additionals::ControllerTest
 
   def setup
     Setting.default_language = 'en'
-    @request    = ActionController::TestRequest.new
-    @response   = ActionController::TestResponse.new
     User.current = nil
   end
 
   def test_get_login_with_welcome_text
-    Setting.plugin_additionals = ActionController::Parameters.new(
-      account_login_bottom: 'Lore impsuum'
-    )
+    with_additionals_settings(account_login_bottom: 'Lore impsuum')
+
     get :login
     assert_response :success
-    assert_template 'login'
 
     assert_select 'input[name=username]'
     assert_select 'input[name=password]'
@@ -24,13 +20,10 @@ class AccountControllerTest < Additionals::ControllerTest
   end
 
   def test_get_login_without_welcome_text
-    Setting.plugin_additionals = ActionController::Parameters.new(
-      account_login_bottom: ''
-    )
+    with_additionals_settings(account_login_bottom: '')
 
     get :login
     assert_response :success
-    assert_template 'login'
 
     assert_select 'input[name=username]'
     assert_select 'input[name=password]'