+++ test/unit/issue_test.rb
@@ -64,12 +64,10 @@ class IssueTest < ActiveSupport::TestCase
   end
 
   def test_auto_assigned_to
-    Setting.plugin_additionals = ActionController::Parameters.new(
-      issue_status_change: '0',
-      issue_auto_assign: '1',
-      issue_auto_assign_status: ['1'],
-      issue_auto_assign_role: '1'
-    )
+    with_additionals_settings(issue_status_change: '0',
+                              issue_auto_assign: '1',
+                              issue_auto_assign_status: ['1'],
+                              issue_auto_assign_role: '1')
 
     issue = Issue.new(project_id: 1, tracker_id: 1, author_id: 3, subject: 'test_create')
     assert issue.save
@@ -77,12 +75,10 @@ class IssueTest < ActiveSupport::TestCase
   end
 
   def test_disabled_auto_assigned_to
-    Setting.plugin_additionals = ActionController::Parameters.new(
-      issue_status_change: '0',
-      issue_auto_assign: '0',
-      issue_auto_assign_status: ['1'],
-      issue_auto_assign_role: '1'
-    )
+    with_additionals_settings(issue_status_change: '0',
+                              issue_auto_assign: '0',
+                              issue_auto_assign_status: ['1'],
+                              issue_auto_assign_role: '1')
 
     issue = Issue.new(project_id: 1, tracker_id: 1, author_id: 3, subject: 'test_create')
     assert issue.save