+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 - updated bootstrap-datepicker to v1.8.0
 - updated d3plus to v2.0.0-alpha.17
 - ruby 2.2.0 is required
+- Redmine 3.4 is required
 
 2.0.12
 ++++++